refactor: simplified search query
