+++ libs/home-cooked/api/recipes/src/lib/home-cooked-api-recipes.controller.ts
@@ -1,19 +1,26 @@
 import { Recipe, RecipeSearch } from '@coderisland/home-cooked/shared/models';
-import { Controller, Get, Param, Query } from '@nestjs/common';
+import { ArgumentMetadata, BadRequestException, Controller, Get, Injectable, Param, ParseIntPipe, PipeTransform, Query } from '@nestjs/common';
 import { map } from 'rxjs/operators';
 import { HomeCookedApiRecipesService } from './home-cooked-api-recipes.service';
+
+@Injectable()
+export class ParseRecipeSearchQuery implements PipeTransform<RecipeSearch> {
+  async transform(value: RecipeSearch, { metatype }: ArgumentMetadata) {
+    const recipeSearch = {
+      ...value,
+      limit: Number(value.limit),
+      page: Number(value.page)
+    };
+    return recipeSearch;
+  }
+}
 
 @Controller('recipes')
 export class HomeCookedApiRecipesController {
   constructor(private homeCookedApiRecipesService: HomeCookedApiRecipesService) {}
 
-  // @Get()
-  // getAll() {
-  //   return this.homeCookedApiRecipesService.getAll();
-  // }
-
   @Get('query')
-  getRecipes(@Query() recipeSearch: RecipeSearch) {
+  getRecipes(@Query(new ParseRecipeSearchQuery()) recipeSearch: RecipeSearch) {
     return this.homeCookedApiRecipesService.getAll(recipeSearch);
   }
