+++ libs/home-cooked/recipes/feature-list/src/lib/shell.component.html
@@ -1,60 +1,15 @@
 <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-  <!-- <form [formGroup]="searchRecipesForm">
-    <input type="text" formControlName="search" />
-    <br />
-    <select formControlName="category" id="categories">
-      <option *ngFor="let cat of categories" [value]="cat">{{ cat }}</option>
-    </select>
-    <br />
-    <input type="text" formControlName="page" />
-    <br />
-    <input type="text" formControlName="limit" />
-  </form> -->
   <div class="text-center">
     <h2 class="text-3xl tracking-tight font-extrabold text-gray-900 sm:text-4xl">Recipes</h2>
     <p class="mt-3 max-w-2xl mx-auto text-xl text-gray-500 sm:mt-4">What do you want to cook today?</p>
   </div>
-  <div>
-    <div>
-      <label for="searchKeyword" class="block text-sm font-medium text-gray-700">Search recipes</label>
-      <div class="mt-1 flex rounded-md shadow-sm">
-        <div class="relative flex items-stretch flex-grow focus-within:z-10">
-          <input
-            type="text"
-            name="searchKeyword"
-            id="searchKeyword"
-            class="focus:ring-indigo-500 focus:border-indigo-500 block w-full rounded-none rounded-l-md sm:text-sm border-gray-300"
-            placeholder="Recipe"
-          />
-          <div class="absolute inset-y-0 right-0 flex items-center">
-            <label for="currency" class="sr-only">Currency</label>
-            <select
-              id="currency"
-              name="currency"
-              class="focus:ring-indigo-500 focus:border-indigo-500 h-full py-0 pl-2 pr-7 border-transparent bg-transparent text-gray-500 sm:text-sm rounded-md"
-            >
-              <option value="">Select category</option>
-              <option>All</option>
-              <option>Beef</option>
-              <option>Pork</option>
-            </select>
-          </div>
-        </div>
-        <button
-          class="-ml-px relative inline-flex items-center space-x-2 px-4 py-2 border border-gray-300 text-sm font-medium rounded-r-md text-gray-700 bg-gray-50 hover:bg-gray-100 focus:outline-none focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500"
-        >
-          <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-            <path
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              stroke-width="2"
-              d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-            ></path>
-          </svg>
-          <span>Search</span>
-        </button>
-      </div>
-    </div>
+  <div class="my-4">
+    <form [formGroup]="searchRecipesForm">
+      <label for="search">
+        Search:
+        <input type="text" formControlName="search" />
+      </label>
+    </form>
   </div>
   <nav class="bg-white py-3 flex items-center justify-between border-t border-gray-200" aria-label="Pagination">
     <div class="hidden sm:block">
