+++ libs/home-cooked/recipes/feature-list/src/lib/shell.component.ts
@@ -18,15 +18,13 @@ import { RecipeSearch } from '@coderisland/home-cooked/shared/models';
 })
 export class ShellComponent implements OnInit {
   searchRecipesForm: FormGroup;
-  categories = ['all', 'beef', 'pork'];
+  categories = ['recipe', 'cuisine', 'ingredients'];
 
   recipes$ = this.store.pipe(select(getAllRecipes));
   constructor(private readonly store: Store<RecipesPartialState>, private readonly formBuilder: FormBuilder) {
     this.searchRecipesForm = this.formBuilder.group({
       search: [''],
       category: [''],
-      page: [1],
-      limit: [10],
     });
   }
 
@@ -34,14 +32,17 @@ export class ShellComponent implements OnInit {
     this.store.dispatch(RecipesActions.init());
 
     this.store.pipe(select(getRecipeSearch)).subscribe((recipeSearch: RecipeSearch) => {
-
       this.searchRecipesForm.patchValue(recipeSearch, { emitEvent: false });
     });
+
     this.searchRecipesForm.valueChanges
       .pipe(
         debounceTime(500),
         distinctUntilChanged(),
-        tap((recipeSearch: RecipeSearch) => this.store.dispatch(RecipesActions.searchRecipes({ recipeSearch }))),
+        tap((recipeSearch: RecipeSearch) => {
+          const data = { ...recipeSearch, limit: 10, page: 1 };
+          this.store.dispatch(RecipesActions.searchRecipes({ recipeSearch: data }));
+        }),
       )
       .subscribe();
   }
