+++ libs/home-cooked/shared/models/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/recipe';
+export * from './lib/recipe-search';
+export * from './lib/paged-result';
