+++ libs/home-cooked/shared/models/src/lib/paged-result.ts
@@ -1,0 +1,7 @@
+import { Recipe } from "./recipe";
+import { RecipeSearch } from "./recipe-search";
+
+export interface PagedResult {
+  recipes: Recipe[];
+  recipeSearch: RecipeSearch;
+}
