+++ libs/home-cooked/shared/models/src/lib/recipe-search.ts
@@ -1,0 +1,6 @@
+export interface RecipeSearch {
+  search: string;
+  category: string;
+  page: number;
+  limit: number;
+}
